% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchEWAS.R
\name{batchEWAS}
\alias{batchEWAS}
\title{Perform batch effect correction}
\usage{
batchEWAS(input,adjustVar = NULL,batch = NULL, plot = TRUE, par.prior = TRUE,
mean.only = FALSE,ref.batch = NULL)
}
\arguments{
\item{input}{An R6 class integrated with all the information.}

\item{adjustVar}{(Optional) Names of the variate of interest and other covariates besides batch,
with each name separated by a comma. Ensure that when correcting for batch effects, the effects
of other factors are appropriately considered and adjusted for.Ensure there are no space. e.g.
"cov1,cov2".}

\item{batch}{Name of the batch variable.}

\item{plot}{(Optional) TRUE give prior plots with black as a kernel estimate of the empirical
batch effect density and red as the parametric.}

\item{par.prior}{(Optional) TRUE indicates parametric adjustments will be used, FALSE indicates
non-parametric adjustments will be used.}

\item{mean.only}{(Optional) Default to FALSE. If TRUE ComBat only corrects the mean of the
batch effect (no scale adjustment).}

\item{ref.batch}{(Optional) NULL If given, will use the selected batch as a reference for batch
adjustment.}
}
\value{
input, An R6 class object integrating all information.
}
\description{
Perform batch effect correction based on the function \code{\link[sva]{ComBat}}
form R package \pkg{sva}. It requires that the "batches" in the data set are known. It uses
either parametric or non-parametric empirical Bayes frameworks for adjusting data for batch effects.
}
\examples{
\dontrun{
res <- initEWAS(outpath = "default")
res <- loadEWAS(input = res, ExpoData = "default", MethyData = "default")
res <- transEWAS(input = res, Vars = "cov1", TypeTo = "factor")
res <- batchEWAS(input = res, batch = "batch", par.prior=TRUE, ref.batch = NULL)
}
}
