% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transEWAS.R
\name{transEWAS}
\alias{transEWAS}
\title{Convert variable type of sample data}
\usage{
transEWAS(input, Vars = "default", TypeTo = "factor")
}
\arguments{
\item{input}{An R6 class integrated with all the information obtained from the loadEWAS function.}

\item{Vars}{Variable names that the user wants to convert types for, with each variable name
separated by a comma. Ensure there are no spaces. e.g. "var1,var2,var3".}

\item{TypeTo}{The type of variable that the function allows to be converted, including numeric and factor.}
}
\value{
input, An R6 class object integrating all information.
}
\description{
Transform the variable types of sample data to the types specified by users.
}
\examples{
\dontrun{
res <- initEWAS(outpath = "default")
res <- loadEWAS(input = res, ExpoData = "default", MethyData = "default")
res <- transEWAS(input = res, Vars = "cov1", TypeTo = "factor")
}
}
