% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootEWAS.R
\name{bootEWAS}
\alias{bootEWAS}
\title{Perform Bootstrap-based Internal Validation}
\usage{
bootEWAS(input, filterP = "PVAL", cutoff = 0.05, CpGs = NULL, times = 500,
bootCI = "perc", filename = "default", seed = NULL)
}
\arguments{
\item{input}{An R6 class integrated with all the information obtained from the startEWAS or
plotEWAS function.}

\item{filterP}{The name of the p value columns such as "PVAL", "FDR", and "Bonfferoni." Users use
this P-value to screen for significance sites and further conduct internal validation.}

\item{cutoff}{The cutoff value of the P-value used to filter for further internal validation.
The default is 0.05.}

\item{CpGs}{The name of the methylation site specified by the user for bootstrap analysis,
separated by commas. Be careful not to have spaces, such as "cpg1,cpg2".}

\item{times}{Number of bootstrap times specified by the user. The default value is 100 times.}

\item{bootCI}{A vector of character strings representing the type of interval to base the test on.
The value should be one of "norm", "basic", "stud", "perc" (the default), and "bca".}

\item{filename}{User-customized .csv file name for storing bootstrap results. If "default", it will
be named as "bootresult".}

\item{seed}{Optional integer seed used for reproducible bootstrap resampling.
If \code{NULL} (default), the current random-number-generator state is used.}
}
\value{
input, An R6 class object integrating all information.
}
\description{
Users can perform internal validation of the identified differently methylated
sites based on the bootstrap method.
}
\examples{
res <- initEWAS(export = FALSE)
res <- loadEWAS(input = res, ExpoData = "default", MethyData = "default")
res <- transEWAS(input = res, Vars = "cov1", TypeTo = "factor")
res$Data$Methy <- res$Data$Methy[seq_len(50), , drop = FALSE]
res <- startEWAS(
  input = res, chipType = NULL, model = "lm",
  expo = "default", adjustP = TRUE, core = 1
)
res <- bootEWAS(input = res, CpGs = res$result$probe[1], times = 5, seed = 1)
}
