% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyEWAS-package.R
\docType{package}
\name{easyEWAS-package}
\alias{easyEWAS}
\alias{easyEWAS-package}
\title{easyEWAS package}
\description{
Tools for conducting epigenome-wide association studies (EWAS) and visualizing results. Users provide sample metadata and methylation matrices to run EWAS with linear models, linear mixed-effects models, or Cox models. The package supports downstream visualization, bootstrap validation, enrichment analysis, batch effect correction, and differentially methylated region (DMR) analysis with optional parallel computing. Methods are described in Wang et al. (2025) \doi{10.1093/bioadv/vbaf026}, Johnson et al. (2007) \doi{10.1093/biostatistics/kxj037}, and Peters et al. (2015) \doi{10.1186/1756-8935-8-6}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ytwangZero/easyEWAS}
  \item \url{https://easyewas-tutorial.github.io/}
  \item Report bugs at \url{https://github.com/ytwangZero/easyEWAS/issues}
}

}
\author{
\strong{Maintainer}: Yuting Wang \email{ytwang@pku.edu.cn}

Authors:
\itemize{
  \item Xu Gao \email{xu.gao@pku.edu.cn}
}

}
\keyword{internal}
