% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initEWAS.R
\name{initEWAS}
\alias{initEWAS}
\title{Initialize the EWAS module}
\usage{
initEWAS(outpath = "default")
}
\arguments{
\item{outpath}{The user-specified path is used to store a generated folder named "EWASresult"
that contains all the analysis results. If "default" is specified, the folder will be generated
in the current working directory.}
}
\value{
input, an R6 class object integrating all information.
}
\description{
This function is designed to generate an R6 class for storing all the data
and results of EWAS analysis, and also to create a folder on the local computer for
storing the analysis results.
}
\examples{
\dontrun{
res <- initEWAS(outpath = "default")
}
}
