% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEWAS.R
\name{plotEWAS}
\alias{plotEWAS}
\title{Visualize the results of EWAS analysis}
\usage{
plotEWAS(input, pval = "PVAL", ...)
}
\arguments{
\item{input}{An R6 class integrated with all the information obtained from the startEWAS function.}

\item{p}{The user needs to specify the name of the p value selected for the result
visualization.}

\item{threshold}{The significant threshold.If threshold = 0 or NULL, then the threshold line will
not be added.}

\item{file}{The format of the output image file, including "jpg","pdf","tiff", and "png".}
}
\value{
input, An R6 class object integrating all information.
}
\description{
Visualize EWAS results based on the CMplot package, including Manhattan
plots, QQ plots, etc. Optional parameters are the same as those in the CMplot function.
}
\examples{
\dontrun{
res <- initEWAS(outpath = "default")
res <- loadEWAS(input = res, ExpoData = "default", MethyData = "default")
res <- transEWAS(input = res, Vars = "cov1", TypeTo = "factor")
res <- startEWAS(input = res, chipType = "EPICV2", model = "lm", expo = "default", adjustP = TRUE)
res <- plotEWAS(input = res, p = "PVAL")
}
}
