% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startEWAS.R
\name{startEWAS}
\alias{startEWAS}
\title{Perform EWAS Analysis}
\usage{
startEWAS(input, filename ="default", model = "lm", expo = "default",
cov = NULL,random = NULL, adjustP = TRUE, core = "default")
}
\arguments{
\item{input}{An R6 class integrated with all the information obtained from the loadEWAS or
transEWAS function.}

\item{filename}{User-customized .csv file name for storing EWAS results. If
"default" is chosen, it will be named as "ewasresult".}

\item{model}{The statistical models used for EWAS analysis include "lm" (general linear
regression), "lmer" (linear mixed-effects model), and "cox" (Cox proportional hazards model).
The default model is "lm".}

\item{expo}{Name of the exposure variable used in the EWAS analysis. The default
is a variable in the example data.}

\item{cov}{Name(s) of covariate(s) used in the EWAS analysis, with each  name separated by
a comma. Ensure there are no space. e.g. "cov1,cov2,cov3".}

\item{random}{Random intercept item name, used only when selecting the "lmer" model.}

\item{time}{When the user selects the Cox proportional risk model, the name
of the time variable needs to be specified.}

\item{status}{When the user selects the Cox proportional risk model, the name
of the status variable needs to be specified.}

\item{adjustP}{Whether to calculate adjusted p-values(FDR and Bonferroni correction). The default
is set to TRUE.}

\item{chipType}{The Illumina chip versions for user measurement of methylation data,
including "27K", 450K ","EPICV1", "EPICV2", and "MSA". The default is "EPICV2".}

\item{core}{The number of cores used during parallel computation. If set to default, it calculates
the maximum number of available physical cores minus 1 and treats this as an operational kernel.}
}
\value{
input, An R6 class object integrating all information.
}
\description{
Perform EWAS analysis to obtain the coefficient value, standard
deviation and significance p value (or adjust p value) of each site.
}
\examples{
\dontrun{
res <- initEWAS(outpath = "default")
res <- loadEWAS(input = res, ExpoData = "default", MethyData = "default")
res <- transEWAS(input = res, Vars = "cov1", TypeTo = "factor")
res <- startEWAS(input = res, filename = "default", chipType = "EPICV2", model = "lm",
expo = "default", adjustP = TRUE, core = "default")
}
}
